Remove claims from rootnav. Fix main navigation. Fix android crash when selecting hotspot. Fix address book and address sharing. Add title to collectables top tabs
