+++ src/features/addressBook/AddressQrScanner.tsx
@@ -1,13 +1,13 @@
 import React, { useCallback } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useNavigation } from '@react-navigation/native'
-import Address from '@helium/address'
 import useHaptic from '@hooks/useHaptic'
 import useAlert from '@hooks/useAlert'
 import QrScanner from '@components/QrScanner'
 import { HomeNavigationProp } from '../home/homeTypes'
 import { useAppStorage } from '../../storage/AppStorageProvider'
 import { parsePaymentLink } from '../../utils/linking'
+import { solAddressIsValid } from '../../utils/accountUtils'
 
 const AddressQrScanner = () => {
   const { triggerNotification } = useHaptic()
@@ -19,7 +19,7 @@ const AddressQrScanner = () => {
   const handleBarCodeScanned = useCallback(
     async (data: string) => {
       // scanned qr is an address string
-      if (Address.isValid(data)) {
+      if (solAddressIsValid(data)) {
         setScannedAddress(data)
         triggerNotification('success')
         navigation.goBack()
@@ -28,7 +28,7 @@ const AddressQrScanner = () => {
 
       // scanned qr is a payment link
       const query = parsePaymentLink(data)
-      if (query?.payee && Address.isValid(query.payee)) {
+      if (query?.payee && solAddressIsValid(query.payee)) {
         setScannedAddress(query.payee)
         triggerNotification('success')
         navigation.goBack()
