+++ src/features/collectables/CollectablesNavigator.tsx
@@ -1,12 +1,11 @@
 import React, { memo } from 'react'
 import {
-  createNativeStackNavigator,
-  NativeStackNavigationOptions,
-} from '@react-navigation/native-stack'
-import CollectionScreen from './CollectionScreen'
-import NftDetailsScreen from './NftDetailsScreen'
+  StackNavigationOptions,
+  createStackNavigator,
+} from '@react-navigation/stack'
+import CollectablesTopTabs from './CollectablesTopTabs'
+import HotspotDetailsScreen from './HotspotDetailsScreen'
 import PaymentScreen from '../payment/PaymentScreen'
-import NftList from './NftList'
 import AddNewContact from '../addressBook/AddNewContact'
 import PaymentQrScanner from '../payment/PaymentQrScanner'
 import AddressQrScanner from '../addressBook/AddressQrScanner'
@@ -14,46 +13,57 @@ import NftMetadataScreen from './NftMetadataScreen'
 import TransferCollectableScreen from './TransferCollectableScreen'
 import AddressBookNavigator from '../addressBook/AddressBookNavigator'
 import TransferCompleteScreen from './TransferCompleteScreen'
+import ClaimRewardsScreen from './ClaimRewardsScreen'
+import ClaimAllRewardsScreen from './ClaimAllRewardsScreen'
+import ClaimingRewardsScreen from './ClaimingRewardsScreen'
+import CollectionScreen from './CollectionScreen'
+import NftDetailsScreen from './NftDetailsScreen'
 
-const CollectablesStack = createNativeStackNavigator()
+const CollectablesStack = createStackNavigator()
 
-const cardPresentation: NativeStackNavigationOptions = {
-  presentation: 'card',
-  animation: 'slide_from_bottom',
+const screenOptions: StackNavigationOptions = {
+  headerShown: false,
 }
 
-const cardFadePresentation: NativeStackNavigationOptions = {
-  presentation: 'card',
-  animation: 'fade',
-}
-
-const modalPresentation: NativeStackNavigationOptions = {
-  presentation: 'modal',
-}
-
-const NFTsNavigator = () => {
+const CollectablesStackScreen = () => {
   return (
-    <CollectablesStack.Navigator
-      screenOptions={{
-        headerShown: false,
-        ...modalPresentation,
-      }}
-    >
-      <CollectablesStack.Screen name="NftsScreen" component={NftList} />
+    <CollectablesStack.Navigator screenOptions={screenOptions}>
+      <CollectablesStack.Screen
+        name="CollectablesTopTab"
+        component={CollectablesTopTabs}
+      />
+      <CollectablesStack.Screen
+        name="HotspotDetailsScreen"
+        component={HotspotDetailsScreen}
+      />
+
+      <CollectablesStack.Screen
+        name="PaymentScreen"
+        component={PaymentScreen}
+      />
+
+      <CollectablesStack.Screen
+        name="ClaimRewardsScreen"
+        component={ClaimRewardsScreen}
+      />
+
+      <CollectablesStack.Screen
+        name="ClaimAllRewardsScreen"
+        component={ClaimAllRewardsScreen}
+      />
+      <CollectablesStack.Screen
+        name="ClaimingRewardsScreen"
+        component={ClaimingRewardsScreen}
+      />
+
       <CollectablesStack.Screen
         name="CollectionScreen"
         component={CollectionScreen}
-        options={cardPresentation}
       />
 
       <CollectablesStack.Screen
         name="NftDetailsScreen"
         component={NftDetailsScreen}
-        options={cardPresentation}
-      />
-      <CollectablesStack.Screen
-        name="PaymentScreen"
-        component={PaymentScreen}
       />
 
       <CollectablesStack.Screen
@@ -73,27 +83,24 @@ const NFTsNavigator = () => {
       <CollectablesStack.Screen
         name="NftMetadataScreen"
         component={NftMetadataScreen}
-        options={cardFadePresentation}
       />
 
       <CollectablesStack.Screen
         name="TransferCollectableScreen"
         component={TransferCollectableScreen}
-        options={cardFadePresentation}
       />
 
       <CollectablesStack.Screen
         name="AddressBookNavigator"
         component={AddressBookNavigator}
-        options={cardPresentation}
       />
 
       <CollectablesStack.Screen
         name="TransferCompleteScreen"
         component={TransferCompleteScreen}
-        options={cardPresentation}
       />
     </CollectablesStack.Navigator>
   )
 }
-export default memo(NFTsNavigator)
+
+export default memo(CollectablesStackScreen)
