+++ src/store/slices/appSlice.ts
@@ -2,14 +2,12 @@ import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 import { Cluster } from '@solana/web3.js'
 
 export type AppState = {
-  showCollectablesTabBar: boolean
   showConnectedWallets: boolean
   showBanner: boolean
   cluster?: Cluster
 }
 
 const initialState: AppState = {
-  showCollectablesTabBar: true,
   showConnectedWallets: false,
   showBanner: true,
 }
@@ -20,9 +18,6 @@ const appSlice = createSlice({
   reducers: {
     setCluster: (state, action: PayloadAction<Cluster>) => {
       state.cluster = action.payload
-    },
-    setCollectablesTabBar: (state, action: PayloadAction<boolean>) => {
-      state.showCollectablesTabBar = action.payload
     },
     toggleConnectedWallets: (state) => {
       state.showConnectedWallets = !state.showConnectedWallets
